feat(sprint10): apply formal Korean email style and add document page tracking with split pane preview
- Rewrite ComposeStep prompts to use 격식체 Korean business email format, removing markdown/numbered citations
- Add page_start/page_end columns to document_chunks for PDF page-level tracking
- Implement extractByPage() in DocumentTextExtractor with PDFBox per-page extraction
- Extend ChunkingService with page-mapping overload that resolves chunk offset to page range
- Extend EvidenceItem with fileName/pageStart/pageEnd and batch join in AnalysisService
- Add DocumentDownloadController for file download and PDF page extraction endpoints
- Add react-pdf PdfViewer component with split pane layout in InquiryAnswerTab
- Improve citation display to show document name and page numbers instead of chunk IDs

+++ backend/app-api/src/main/resources/db/migration/V6__retrieval_evidence.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS retrieval_evidence (
     id UUID PRIMARY KEY,
     inquiry_id UUID NOT NULL,
     chunk_id UUID NOT NULL,
-    score DOUBLE NOT NULL,
+    score DOUBLE PRECISION NOT NULL,
     rank_order INT NOT NULL,
     question VARCHAR(4000) NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL,
